wav2vec2_xlsr_pkg/wav2vec2_xlsr_pkg/requirements.txt
jupyter
numpy
pandas
torch
torchaudio
transformers
datasets
tqdm
jiwer
soundfile
pyyaml
